+++ HISTORY.rst
@@ -5,16 +5,19 @@ History
 1.7.1 (XXXX-XX-XX) Support for REST API based communicators, running integrations as services, and connecting to remote integration services
 ------------------
 
-* Added option to return printStatus string via return_str keyword
+* Added option to return printStatus string for YggClass subclasses via return_str keyword
 * Added classes for managing models as services via Flask or RabbitMQ
 * Transitioned from as_function to complete_partial keyword in parse_yaml that can be used for services as well as functions
 * Added client side instrumentation for connecting to remote model services and tests
 * Added support for registering integrations that can be called locally as services
+* Added 'integration-service-manager' CLI for managing service managers
 
 TODO:
 ~~~~~
 
-* Add keyword to specify what type of connection should be used to connect to services
+* Add service configuration options (default commtype, exposed ports, etc.)
+* Add test for connecting to service via import as function
+* Add test for server/wrapped function service
 
 
 1.7.0 (2021-08-26) Support for MPI communicators, MPI execution, and pika >= 1.0.0